bit-string: Fix itty_bit_string_condense and itty_bit_string_transpose
When transposing bit strings we really need to ignore leading zeros
since they are not significant, and tranposing a list shouldn't end
up with a huge number of all zero bit strings in the front of it.

This commit adds a new method itty_bit_string_get_length that returns
the number of actually significant bits, and a similar function for
itty_bit_string_list_t.

It then fixes itty_bit_string_transpose and itty_bit_string_condense
to use those functions.