bit-string: Add mutability enum
Right now we have this weird situation where bit strings may be
wrapping data that is heap allocated or coming from a memory mapped
file.

This means we don't know whether or not to free the data when freeing
the bit string.

This commit adds a new type `itty_bit_string_mutability_t`
to specify whether or not the data is read-only (and also whether or
not it should get freed).

Adapting the code to use this enum demonstrated that
`itty_bit_string_duplicate` isn't a good fit for the api, and it's
unused thus far, so this commit drops it.